requirements.txt
streamlit
pandas
yfinance
finnhub-python
pandas-ta
scikit-learn
nltk
python-dotenv
alpaca-py
plotly
